requirements.txt
streamlit>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
typing>=3.7.4
